+++ src/pages/viewer/Calcs.vue
@@ -225,7 +225,7 @@ export default {
       }
     },
     active (val) {
-      let currentItem = this.itemsCollection[val] || null
+      let currentItem = this.itemsCollection[val] || {}
       if (val) {
         this.$q.localStorage.set('calcs', val)
         this.$router.push(`/calcs/${val}`)