+++ src/pages/viewer/Containers.vue
@@ -192,7 +192,7 @@ export default {
       }
     },
     active (val) {
-      let currentItem = this.itemsCollection[val] || null
+      let currentItem = this.itemsCollection[val] || {}
       if (val) {
         this.$q.localStorage.set('containers', val)
         this.$router.push(`/containers/${val}`)