+++ src/pages/viewer/Streams.vue
@@ -230,7 +230,7 @@ export default {
       }
     },
     active (val) {
-      let currentItem = this.itemsCollection[val] || null
+      let currentItem = this.itemsCollection[val] || {}
       if (val) {
         this.$q.localStorage.set('streams', val)
         this.$router.push(`/streams/${val}`)